requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
Brotli==1.1.0
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dj-database-url==2.3.0
Django==5.1.5
django-celery-results==2.5.1
django-ckeditor==6.7.2
django-js-asset==3.1.2
et_xmlfile==2.0.0
fonttools==4.56.0
gunicorn==23.0.0
h11==0.14.0
kiwisolver==1.4.8
kombu==5.4.2
matplotlib==3.10.0
numpy==2.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
pyxlsb==1.0.10
redis==5.2.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uvicorn==0.34.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
